requirements.txt
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
jsonschema==4.3.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
openapi==1.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
requests==2.27.1
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==2.0.2

